Update SimpleJpaRepository.java
Always that a delete a entity thats not maneged bt JPA, I get ORA-01400: cannot insert NULL. It´s happen because JPA try to insert the entities that is not maneged.
It´s usual to get a couple from config tables and display on screen where user can populate more data or delete, and it my populate or delete a final table to feed others systems. So it can happen to user delete a data that been recovery by Jpa Entity and it will not be in managed by JPA. When a data is not managed by the JPA it all ways try to make a insert (before call deleting), but for delete I can use only the PK, so others non nullable information will not be in the Entity.
It makes more sense to not merge deleting info. Just ignore.